backend/requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
pandas
openpyxl
